add images as background, use default theme

+++ lib/models/onecallapi/forecast.dart
@@ -62,7 +62,7 @@ class Forecast {
       temp: json['current']['temp'].toDouble(),
       condition: Mapping.mapStringToWeatherCondition(weather['main']),
       description: weather['description'],
-      feelLikeTemp: json['current']['feels_like'],
+      feelLikeTemp: json['current']['feels_like'].toDouble(),
       date: date,
     );
